moved utils and config into own folders
