+++ packages/modular-scripts/DO_NOT_IMPORT_THIS_OR_YOU_WILL_BE_FIRED_craco.config.js
@@ -148,7 +148,7 @@ module.exports = {
           },
         },
         setupFiles: jestConfig.setupFiles
-          .concat([path.join(__dirname, './jest-setupEnvironment.js')])
+          .concat([require.resolve('react-scripts/config/env.js')])
           .concat(
             glob.sync(
               `${absoluteModularGlobalConfigsPath}/setupEnvironment.{js,ts,tsx}`,
