+++ packages/modular-scripts/src/buildPackage.ts
@@ -29,8 +29,8 @@ import * as ts from 'typescript';
 import * as fse from 'fs-extra';
 
 import builtinModules from 'builtin-modules';
-import getModularRoot from './getModularRoot';
-import { getAllWorkspaces } from './workspaces';
+import getModularRoot from './utils/getModularRoot';
+import { getAllWorkspaces } from './utils/workspaces';
 
 const modularRoot = getModularRoot();
