+++ packages/modular-scripts/src/config/jest/index.ts
@@ -1,0 +1,3 @@
+import { jestConfig } from '../index';
+
+export default jestConfig;
