+++ packages/modular-scripts/src/start.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
-import getModularRoot from './getModularRoot';
-import isModularType from './isModularType';
-import execSync from './execSync';
+import getModularRoot from './utils/getModularRoot';
+import isModularType from './utils/isModularType';
+import execSync from './utils/execSync';
 
 import { cracoBin, packagesRoot, cracoConfig } from './config';
