+++ packages/modular-scripts/src/utils/execSync.ts
