+++ packages/modular-scripts/src/utils/getModularRoot.ts
