+++ packages/modular-scripts/src/utils/isModularType.ts
