+++ packages/modular-scripts/src/utils/workspaces.ts
